[REFACTOR] import as, duplicate style code
environment
