--- src/components/EnvironmentHeader/style.tsx
@@ -1,26 +1,0 @@
-import styled from "styled-components";
-
-export const EnvironmentGroup = styled.div`
-    width: 10%;
-    font-weight: bold;
-`
-
-export const PlatformGroup = styled.div`
-    width: 15%;
-    font-weight: bold;
-`
-
-export const ClientDomainGroup = styled.div`
-    width: 30%;
-    font-weight: bold;
-`
-
-export const ServerDomainGroup = styled.div`
-    width: 30%;
-    font-weight: bold;
-`
-
-export const ButtonGroup = styled.div`
-    float: right;
-    gap: 10px;
-`
